+++ src/cpu/c++/lapack.cpp
@@ -62,7 +62,7 @@ namespace lahva
             };
             
             LPCK_INT n = a.shape().first;
-            if (ipiv.size() != n)
+            if (static_cast<size_t>(ipiv.size()) != static_cast<size_t>(n))
             {
                 ipiv = Vector<LPCK_INT>(n);
             }
@@ -187,7 +187,7 @@ namespace lahva
             LPCK_INT n = a.shape().first;
             LPCK_INT nrhs = b.shape().second;
             CholeskyFactorization(a);
-            if (b.shape().first != n)
+            if (static_cast<size_t>(b.shape().first) != static_cast<size_t>(n))
             {
                 throw std::runtime_error("B should have as much rows as A. B(n, nrhs)");
             }
@@ -255,7 +255,7 @@ namespace lahva
             LPCK_INT nrhs = b.shape().second;
             Vector<LPCK_INT> ipiv(n);
             LUFactorization(a, ipiv);
-            if (b.shape().first != n)
+            if (static_cast<size_t>(b.shape().first) != static_cast<size_t>(n))
             {
                 throw std::runtime_error("B should have as much rows as A. B(n, nrhs)");
             }
@@ -292,7 +292,7 @@ namespace lahva
             LPCK_INT nrhs = b.shape().second;
             Vector<LPCK_INT> ipiv(n);
             LUFactorization(a, ipiv);
-            if (b.shape().first != n)
+            if (static_cast<size_t>(b.shape().first) != static_cast<size_t>(n))
             {
                 throw std::runtime_error("B should have as much rows as A. B(n, nrhs)");
             }
@@ -328,7 +328,7 @@ namespace lahva
             LPCK_INT nrhs = b.shape().second;
             Vector<LPCK_INT> ipiv(n);
             LUFactorization(a, ipiv);
-            if (b.shape().first != n)
+            if (static_cast<size_t>(b.shape().first) != static_cast<size_t>(n))
             {
                 throw std::runtime_error("B should have as much rows as A. B(n, nrhs)");
             }
@@ -365,7 +365,7 @@ namespace lahva
             };
             LPCK_INT n = a.shape().first;
             LPCK_INT nrhs = b.shape().second;
-            if (b.shape().first != n)
+            if (static_cast<size_t>(b.shape().first) != static_cast<size_t>(n))
             {
                 throw std::runtime_error("B should have as much rows as A. B(n, nrhs)");
             }
@@ -398,7 +398,7 @@ namespace lahva
             LPCK_INT info = 0;
             LPCK_INT n = a.shape().first;
             LPCK_INT nrhs = b.shape().second;
-            if (b.shape().first != n)
+            if (static_cast<size_t>(b.shape().first) != static_cast<size_t>(n))
             {
                 throw std::runtime_error("B should have as much rows as A. B(n, nrhs)");
             }
@@ -431,7 +431,7 @@ namespace lahva
             LPCK_INT info = 0;
             LPCK_INT n = a.shape().first;
             LPCK_INT nrhs = b.shape().second;
-            if (b.shape().first != n)
+            if (static_cast<size_t>(b.shape().first) != static_cast<size_t>(n))
             {
                 throw std::runtime_error("B should have as much rows as A. B(n, nrhs)");
             }
@@ -464,7 +464,7 @@ namespace lahva
             LPCK_INT info = 0;
             LPCK_INT n = a.shape().first;
             LPCK_INT nrhs = b.shape().second;
-            if (b.shape().first != n)
+            if (static_cast<size_t>(b.shape().first) != static_cast<size_t>(n))
             {
                 throw std::runtime_error("B should have as much rows as A. B(n, nrhs)");
             }
@@ -497,7 +497,7 @@ namespace lahva
             LPCK_INT info = 0;
             LPCK_INT n = a.shape().first;
             LPCK_INT nrhs = b.shape().second;
-            if (b.shape().first != n)
+            if (static_cast<size_t>(b.shape().first) != static_cast<size_t>(n))
             {
                 throw std::runtime_error("B should have as much rows as A. B(n, nrhs)");
             }
@@ -542,7 +542,7 @@ namespace lahva
             LPCK_INT info = 0;
             LPCK_INT n = a.shape().first;
             LPCK_INT nrhs = b.shape().second;
-            if (b.shape().first != n)
+            if (static_cast<size_t>(b.shape().first) != static_cast<size_t>(n))
             {
                 throw std::runtime_error("B should have as much rows as A. B(n, nrhs)");
             }
